add  the function of deleting file

+++ disk.cpp
@@ -4,8 +4,6 @@
 
 Disk::Disk()
 {
-//    diskFile.open("D:\\QtWorkspace\\FileSystemSimulator\\disk.vdisk",
-//                  std::ios::in|std::ios::out|std::ios_base::binary);
     diskFile.open("disk.vdisk",
                   std::ios::in|std::ios::out|std::ios_base::binary);
 }
@@ -14,7 +12,7 @@ Disk::~Disk(){
     diskFile.close();
 }
 
-Block Disk::readBlock(unsigned index){
+Block Disk::readBlock(uint8_t index){
     //此处应检查index是否小于127
     //使用异常？静态断言？或其它？
     uint8_t block[64];
@@ -31,7 +29,4 @@ void Disk::writeBlock(Block block, uint8_t index){
     diskFile.seekp(index*64,std::ios::beg);
     diskFile.write((char*)block.data(),64);
     diskFile.flush();
-//    diskFile.close();
-//    diskFile.open("D:\\QtWorkspace\\FileSystemSimulator\\disk.vdisk",
-//                  std::ios::in|std::ios::out|std::ios_base::binary);
 }
